moved timer component and  set timer in waiting
